Correctly detect win/lose state
If the game is won or lost this will be detected,
the game updated and the outcome shown in UI.

Opening root creates a new game.

The board has been wrapped in a turbo frame for more
efficient updating and the update endpoint directly
renders and returns the page. Before, clicking involved
two trips to the server to update the board.