Ensure correct rendering in case of concurrent clicks
The way this is ensured is by always fetching the list
from the database. Since we're ordering by id the database
ensures the ids are always increasing and whenever we fetch
all clicks so far we know it's not possible for any other
process to insert new clicks in the middle of the list, only
at the end.

This means that the database guarantees us that we always
see a correct state of the game at some point in time.
We might see it stale some clicks were recorded after we
loaded our game object but the game object will always be
a valid current or past state of the game.

Due to how SQLite works (by database essentially being
in the memory of the same process as the app) this is
introducing almost no additional performance penalty.