Improve the page a bit
Still not award winning design but the front page now
shows the explanation of what this is and who made it.

It also shows the total score of Mines vs Humanity.
The score is cached on the current game id. Since there is only
a maximum of one game being played at any time, the score can't
change until the currently played game id changed.