Add mine location flagging to the gameplay
The core is logic is very simple and implemented in the
Minesweeper::Game board. The rest of the application just
deals with storing it in the database and providing a UI.

The UI part is unfinished here. In theory the `contextmenu`
event is perfect for this. And it is on desktop but on
iOS 13+ it's no longer firing on long click. So this doesn't
work on iOS.

A future improvement is needed to patch up the iOS UI.
I have a feeling that patching this up may end up being
more code than the whole feature is now.