Do not show clicks count on games history
The number of games has exploded faster than I expected which is a performance issue on the history page. A quick fix is to remove click counts from it. Since I'm on vacation and doing this on a tablet through github Web editor,  this will have to do for now.