Optimise flooding algorithm
The initial algorithm of recursively calling
reveal method is pretty simple buy very inefficient.

The recursion has an overhead and it easily hits
a "Stack level too deep" error on any board larger
than the regular expert level.

Like any simple recursion it can be easily unrolled
with help of a stack and this is what's done here.
The peformance increases by orders of magnitude.
I can't even measure it properly since boards for which
the new approach calculates flooding in a few ms would
hit the stack level too deep on the recursive approach.